app/requirements.txt
azure-storage-blob
prometheus_client
pyYAML